refactor: CammelCase for folder structure
