+++ microsite-next/src/components/bannerSection/bannerSection.module.scss
