+++ microsite-next/src/components/bannerSection/bannerSectionGrid.module.scss
