+++ microsite-next/src/components/contentBlock/contentBlock.module.scss
