+++ microsite-next/src/pages/home/_home.tsx
@@ -1,15 +1,15 @@
 import BrowserOnly from '@docusaurus/BrowserOnly';
 import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import { BannerSection } from '@site/src/components/banner-section/banner-section';
-import { BannerSectionGrid } from '@site/src/components/banner-section/banner-section-grid';
-import { ContentBlock } from '@site/src/components/content-block/content-block';
+import { BannerSection } from '@site/src/components/bannerSection/bannerSection';
+import { BannerSectionGrid } from '@site/src/components/bannerSection/bannerSectionGrid';
+import { ContentBlock } from '@site/src/components/contentBlock/contentBlock';
 import Layout from '@theme/Layout';
 import { clsx } from 'clsx';
 import React, { useState } from 'react';
 
 import homeStyles from './home.module.scss';
-import { HubSpotNewAdoptersForm } from './_hub-spot-new-adopters-form';
+import { HubSpotNewAdoptersForm } from './_hubSpotNewAdoptersForm';
 
 const hiddenNewsletterBannerKey = 'hiddenNewsletterBanner';
