+++ microsite-next/src/pages/plugins/index.tsx
@@ -3,7 +3,7 @@ import Layout from '@theme/Layout';
 import clsx from 'clsx';
 import React from 'react';
 
-import { IPluginData, PluginCard } from './_plugin-card';
+import { IPluginData, PluginCard } from './_pluginCard';
 import pluginsStyles from './plugins.module.scss';
 
 //#region Plugin data import
